Add data and read input

+++ fenwick.cpp
@@ -1,10 +1,10 @@
 #include "fenwick.hpp"
 
 
-fenwick::fenwick(const long int input_size) :
+fenwick::fenwick(const long long int input_size) :
     _size(input_size)
 {
-    _tree = new long int[_size];
+    _tree = new long long int[_size];
 }
 
 fenwick::~fenwick()
@@ -12,9 +12,9 @@ fenwick::~fenwick()
     delete [] _tree;
 }
 
-long int fenwick::read(long int idx)
+long long int fenwick::read(long long int idx)
 {
-    long int sum = 0;
+    long long int sum = 0;
     while (idx > 0)
     {
         sum += _tree[idx];
@@ -23,7 +23,7 @@ long int fenwick::read(long int idx)
     return sum;
 }
 
-void fenwick::update(long int idx , const long int val)
+void fenwick::update(long long int idx , const long long int val)
 {
     while (idx <= _size)
     {
